work through issues with Path.upload
turns out that the blackfynn api delete operation is non-blocking
which induces all sorts of fun issues (file under WAT)

also my fiddling of the upload_files function seems to have broken
blackfynn's ability to generate hashes for those packages, I thought I
had fixed that, maybe I didn't, in which case it needs to be fixed now

another issue is that somewhere during the update process the cache
metadata is getting klobbered, I dealt with the obvious cases, but
somehow it is still getting overwritten so my hack around missing
checksums doesn't work, since impossible to compute the checksum on
a package that has already been deleted (and that has not been cached)

I think I know how to fix this, once we get the package id after
upload, we need to immediately add a copy to the local object store,
a bit of additional functionliaty will be to prevent upload if the
local file has not changed from the previous object. If the local
object store is deleted then yes we are still in trouble but at that
point we will want that error to happen.