convert OntId to OntTerm during term expansion and rdf conversion
Since we are in the process of moving all network related activity
related to identifiers to its own phase these parts of the pipeline
are now receiving uninstrumented versions of identifiers which they
now need to resolve.

This fixes the bug where organs, techniques, modalities, etc. were not
being expanded in the json export.