viewer major performance improvements
Render a jview for each dataset json so that we don't have to
recompute it. Have to swap parent with an invisible frame, but that is
fairly straight forward. In combination with incremental population of
the hierlist when nodes are opened it is about as responsive as we are
likely to get. Unfortunately show all nodes seems to tank performance,
probably due to the errors json, but worth investigating.