sparcur.simple.utils.fetch_paths_parallel switch to Async/deferred
You know that crazy stochastic file size mismatch error? We have it
cornered now. Nearly have the full root cause found as well, but not
quite, there are still some weird things going on.

For example, if I am using multiprocessing to retrieve different
files, then shouldn't nothing be shared and shouldn't it be impossible
for the url for one file to be returned from a requests session.get
that specified the other? I kind of want to check that the request url
matches across a single call to session.get because this is so insane.

In short, the reason why a mismatched file size occures is becuase a
call to BlacckfynnRemote.get_file_url -> utils.ApiWrapper.get_file_url
can seeminly at random return a url pointing to the wrong file!?! On
examination you will notice that that url returned points to THE WRONG
FILE, but also usually a file was also requested close in time to the
one that got the wrong url.

Looking over the logs again, it is clear to me now that the issue is
not that there is a race condition in any of the joblib code, because
it is not that we are somehow reading the results of session.get
twice.  The reason we know this is that rest of the long url with the
X-Amazon fields is different (eternal screaming).

Switching to Async/deferred avoids the issue, probably by going more
slowly, and by not sending a multiple requests during the same
microsecond, which if I had to bet would be why these things get
confused on the remote side.

I am going to test with this configuration to see whether we encounter
similar or other issues.

The issue was found when I noticed that the mismatched actual size of
a subjects.xlsx file was the exact size of the samples.xlsx file.
Which is the kind of utterly terrifying thing that you hope to never
see because it means that something has gone Mickensly wrong.