Create split_excel_sheet.py
This script can be used to recursively read excel files and for each sheet in the file, create a new excel file. 

mention the source in the root_directory  and the destination in the new_directory.  The script reads any excel sheet recursively in the folder and following the same folder structure saves the new excel files those are generated. 

The error_log.txt file contains the logs of file names in which any type of error occurs.