viewer improvements and bugfixes
the dataset list view includes dataset owners, and updated datetime
it can also be sorted by clicking on the column headers

the upgrade should go smoothly from cache since we now append any
additional columns needed beyond those found in the cache

sparcur.simple.utils for-racket has been updated to produce owners and
updated columns, organization members are now retrieved and cached in
memory on first request so there are 2 requests that go out, but since
refreshing dataset metadata has been made to work async in a separate
thread and correctly restore state afterward the slowness isn't as
much of an issue as it is on first run with no cache

more buttons are now correctly disabled when their action would fail
e.g. open folder if nothing has been fetched, json when nothing has
been exported etc. there are some that might have slipped through the
cracks, but in general these should be much more readable

fetch+export now reports when the fetch half has finished and updates
the ui buttons accordingly

a number of symbols that used to be parameters are now case-lambdas
because there were race conditions related to the thread-local
behavior of parameters that produced crazy out of order events that
would e.g. result in multiple jviews, switching between
current-datasets and current-datasets-view when sorting etc. the
general principle seems to be that if it is an internal state variable
that follows the gui then it should probably not be a parameter

also fixed issue where lead/trail whitespace would break search

viewer now depends on gregor to render dates in the local timezone

fixed issue where the config cache path and log path did not match the
default, now back to including racket and datasets folder as expected

fix c:l keybind -> c:k to go to the dataset search box (match browser)

dataset search now happens off the main thread and is synchornized
back by sleeping for +100+ 150 ms (100 ms seems too short and errors
and races can and do happen) and seeing if the input text has changed,
that is sufficiently long to avoid issues with ordering and but also
perceptually quite short so the user almost never notices, the general
principle seems to be that if you don't want to deal with locking
etc. you can fake it by sleeping in the thread and the re-reading the
variable that was passed in at call time to see if it has changed, and
if it has then you no longer have priority and are done