sparcron fix issue with missing rmeta cache
in the logic for sparcron fetch could be false when rmeta was not
present either because the cache had been removed or a previous run
had failed for some reason but the local file system was up-to-date,
e.g. because some other process had pulled the fs but not rmeta

we now check whether rmeta is present and fetch it it is missing

in sparcron we also use mset to set all variables at the same time in
redis at the end of ret_val_exp

two new internal versions have been added for rmeta version and fs
translation version, they are used to detect whether there has been a
change in the expected structure of fs and rmeta so that pipelines and
be rerun if those values change e.g. because we change how we handle
packages with multiple files, the fs version is embedded in the
dataset folder xattrs under user.sparcur.fsversion, the implementaiton
requires that the whole dataset hierarchy be pulled, sub directory
sync is not supported atm the value for fs version is embedded in prov
and the schema has been updated accordintly, rmeta version is embedded
in the cached rmeta blob itself under the _meta_version key

pipelines.JSONPipeline.adds is now processed by passing the instance
of the pipeline that is running to the lambda in adds instead of the
lifters, this is still not optimal but should help simplify a future
transition to something that is not so utterly broken