participants.org add material object to hierarchy
it was not clear from the existing hierarchy where non-biological
material objects would go in the hierarchy, most of them would go in
as subjects, e.g. if one were to put a can in an mri scanner, there
might be some confuion about where to put physical samples, such as
rock samples or ice core samples because the "subject" would
technically be "the whole planet earth" and you would want to record
the location from which the sample was collected, etc. further
guidance and refinement is needed here