viewer first pass at surfacing errors from path-metadata.json
will likely need another pass or two, but with this in place we can
surface issues with e.g. file names, unknown mimetypes, zero byte
files etc.

this is a workaround to make this visible to curators, errors in path
metadata are not currently counted in dataset metadata totals and are
not integrated into that framework, ultimately they should be
integrated into that framework and should be derived from the single
file (objects) pipelines as we move toward replacing the existing
pipelines with ones based on the single file pipelines