paths path-metadata.json include compacted error report
a starting point for surfacing issues with paths to curation, this
needs a proper design pass, we want to move away from the combined
path-metadata.json and to the single file index, but we also need a
way to run e.g. json schema constraints in an sds version dependent
manner, etc. so this is a placeholder specifically to surface things
that curation needs