viewer better handling for python first time case
these changes fill the remining gaps from the previous commit to allow
the viewer to operate without a full configuration without emitting
errors in the process

also a stub to handle cases where xdg-open is not on a system, which
should be rare, but happens e.g. in our docker images where we don't
actually install the user facing side of x11 because the x server is
expected to be running on the host system, possibly need to revisit
that decision, but we don't e.g. put firefox in the dev environment